requirements.txt.txt
streamlit
Pillow==9.0.0
tensorflow-cpu
numpy
pytest-shutil
scikit-learn
tensorflow